[one-cmds] Introduce Command.py
This commit introduces Command.py to one-cmds.
